+++ docs/additional_information/changelog.rst
@@ -10,6 +10,12 @@ Added
 ^^^^^
 
 - Option to save the anomaly scores within a workflow.
+- Implement ``LocalPolynomialApproximation`` anomaly detector.
+- Implement ``UCRScore`` evaluation metric.
+- Implement ``ChronosAnomalyDetector`` model, with optional dependency ``chronos``
+- Implement ``MOMENTAnomalyDetector`` model, with optional dependency ``moment``
+- Implement ``TimeMoEAnomalyDetector`` model, with optional dependency ``time_moe``
+- Implement ``RobustRandomCutForestAnomalyDetector`` model
 
 Changed
 ^^^^^^^
@@ -21,7 +27,7 @@ Fixed
   This was especially necessary for time indexes in a dataset.
 
 [0.4.2] - 2025-07-03
-------
+--------------------
 
 .. warning::
    We ended the support for Python version 3.9